requirements.txt
fastapi
uvicorn
google-play-scraper
python-dotenv
google-genai
openai
requests
aiofiles
python-multipart
